requirements.txt
alembic==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-MySQLdb==0.2.0
Flask-pymysql==0.2.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
pyepayco==1.0.0
pylast==4.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
